Refactor reportv4 to report packages
