+++ safe/gui/analysis_utilities.py
@@ -12,9 +12,8 @@ from safe.definitions.report import (
     standard_impact_report_metadata_pdf,
     report_a4_blue)
 from safe.impact_function.style import hazard_class_style
-from safe.gui.tools.minimum_needs.needs_profile import NeedsProfile
-from safe.reportv4.report_metadata import ReportMetadata
-from safe.reportv4.impact_report import ImpactReport as ImpactReportV4
+from safe.report.report_metadata import ReportMetadata
+from safe.report.impact_report import ImpactReport
 
 __copyright__ = "Copyright 2016, The InaSAFE Project"
 __license__ = "GPL version 3"
@@ -35,7 +34,7 @@ def generate_impact_report(impact_function, iface):
     # create impact report instance
     report_metadata = ReportMetadata(
         metadata_dict=standard_impact_report_metadata_pdf)
-    impact_report = ImpactReportV4(
+    impact_report = ImpactReport(
         iface,
         report_metadata,
         impact_function=impact_function)
@@ -68,7 +67,7 @@ def generate_impact_map_report(impact_function, iface):
     # create impact report instance
     report_metadata = ReportMetadata(
         metadata_dict=report_a4_blue)
-    impact_report = ImpactReportV4(
+    impact_report = ImpactReport(
         iface,
         report_metadata,
         impact_function=impact_function)
