+++ safe/report/extractors/analysis_result.py
@@ -10,7 +10,7 @@ from safe.definitions.hazard_category import (
     hazard_category_single_event,
     hazard_category_multiple_event)
 from safe.definitions.hazard_classifications import all_hazard_classes
-from safe.reportv4.extractors.util import layer_definition_type, \
+from safe.report.extractors.util import layer_definition_type, \
     round_affecter_number
 from safe.utilities.i18n import tr
 
@@ -25,11 +25,11 @@ def analysis_result_extractor(impact_report, component_metadata):
 
     :param impact_report: the impact report that acts as a proxy to fetch
         all the data that extractor needed
-    :type impact_report: safe.reportv4.impact_report.ImpactReport
+    :type impact_report: safe.report.impact_report.ImpactReport
 
     :param component_metadata: the component metadata. Used to obtain
         information about the component we want to render
-    :type component_metadata: safe.reportv4.report_metadata.
+    :type component_metadata: safe.report.report_metadata.
         ReportComponentsMetadata
 
     :return: context for rendering phase
