+++ safe/report/extractors/infographic_elements/__init__.py
