+++ safe/report/extractors/infographic_elements/svg_charts.py
