+++ safe/report/test/fixtures/analysis_sample/1-exposure.geojson
