+++ safe/report/test/fixtures/analysis_sample/14-impact.geojson
