+++ safe/report/test/fixtures/analysis_sample/17-aggregation.geojson
