+++ safe/report/test/fixtures/analysis_sample/18-analysis.geojson
