+++ safe/report/test/fixtures/analysis_sample/2-hazard.xml
