+++ safe/report/test/fixtures/controls/impact-function-impact-report-output-2.html
